appliance: restrict ext2 image to 512M
The 4G default is obviously an overkill for our usage.

Arch Linux produce > 256M < 512M, let's set the limit to 512M
supermin packages by package list, not by finding library deps with ldd
(like mkinitcpio), and that why arch becomes so big (by pulling gcc-libs)

We won't need much outside from a statically-linked busybox. The arch
situation is from kmod, which we need for compressed modules (busybox
seemless decompression is able to handle xz, but i'm not sure if it's
enabled on arch, and we will likely still need kmod for zstd in the
future). It is a special case for a distro that is conserative on
spliting packages. For other distros the actual use of space in the
image should be very small.