c2v, appliance: use and bundle our own setuidgid
We aim to mimic contianer environment, but busybox setuidgid does not
support supplementary groups, which OCI spec requires when group is not
specified. Also, busybox on Debian and alike lacks this applet.

The build system is seperated and a prebuilt binary is provided because
to use it in the last stage, static linking is way cleaner, and we
resolve passwd and groups at run-time, which, on glibc, still needs
specific version of the lib present for the sake of NSS.

OCI spec requires resolving by traditional files only. We don't need
NSS. So the best way is to static link with other static-link-friendly
libc, and we bundle a musl-linked one.

TODO: Consider moving uid/gids resolving part to c2v stage, to enable
use of glibc and remove the need of bundling, but note that we still
need it to be statically linked for simplicity and cleanliness.