v2c: fix hardlink detection
before this, nlinks passed to libarchive were always 0, making linkify
consider all entries.
this fixes it, but on btrfs snapshots, (ino, dev) are the same without
bumping nlink because they are not really hardlinks and are mistaken
before this. as a result, this now make result from snapshotted image 2x
large, emphasizing the need of detecting (and opt-in ignoring) snapshots